+++ tests/nvmeof/workflows/ha.py
@@ -15,8 +15,10 @@ from ceph.parallel import parallel
 from ceph.utils import get_node_by_id, get_openstack_driver
 from ceph.waiter import WaitUntil
 from tests.nvmeof.workflows.initiator import NVMeInitiator, validate_initiator
+from tests.nvmeof.workflows.nvme import NVMeService
 from tests.nvmeof.workflows.nvme_gateway import NVMeGateway
 from tests.nvmeof.workflows.nvme_utils import deploy_nvme_service
+from tests.nvmeof.workflows.utils import get_current_timestamp, string_to_dict
 from utility.log import Log
 from utility.retry import retry
 from utility.utils import log_json_dump
@@ -24,54 +26,30 @@ from utility.utils import log_json_dump
 LOG = Log(__name__)
 
 
-def get_current_timestamp():
-    return time.perf_counter(), time.asctime()
-
-
-class HighAvailability:
+class HighAvailability(NVMeService):
     def __init__(self, ceph_cluster, gateways, **config):
         """Initialize NVMeoF Gateway High Availability class.
 
         Args:
-            cluster: Ceph cluster
+            ceph_cluster: Ceph cluster
             gateways: Gateway node Ids
             config: HA config
         """
-        self.cluster = ceph_cluster
-        self.config = config
-        self.gateways = []
-        self.mtls = config.get("mtls")
-        self.gateway_group = config.get("gw_group", "")
-        self.orch = Orch(cluster=self.cluster, **{})
-        self.daemon = Daemon(cluster=self.cluster, **{})
-        self.host = Host(cluster=self.cluster, **{})
-        self.nvme_pool = config["rbd_pool"]
+        super().__init__(config=config, ceph_nodes=gateways, ceph_cluster=ceph_cluster)
+
+        self.orch = Orch(cluster=self.ceph_cluster)
+        self.daemon = Daemon(cluster=self.ceph_cluster)
+        self.host = Host(cluster=self.ceph_cluster)
         self.clients = []
         self.initiators = {}
 
-        for gateway in gateways:
-            gw_node = get_node_by_id(self.cluster, gateway)
-            self.gateways.append(NVMeGateway(gw_node, self.mtls))
-
-        self.ana_ids = [i.ana_group_id for i in self.gateways]
+        self.ana_ids = [gw.ana_group_id for gw in self.gateways]
         self.fail_ops = {
             "systemctl": self.system_control,
             "daemon": self.ceph_daemon,
             "power_on_off": self.power_on_off,
             "maintanence_mode": self.maintanence_mode,
         }
-
-    def check_gateway(self, node_id):
-        """Check node is NVMeoF Gateway node.
-
-        Args:
-            node_id: Ceph node Id (ex., node6)
-        """
-        for gw in self.gateways:
-            if gw.node.id == node_id:
-                LOG.info(f"[{node_id}] {gw.node.hostname} is NVMeoF Gateway node.")
-                return gw
-        raise Exception(f"{node_id} doesn't match to any gateways provided...")
 
     def get_or_create_initiator(self, node_id, nqn):
         """Get existing NVMeInitiator or create a new one for each (node_id, nqn)."""
@@ -82,33 +60,6 @@ class HighAvailability:
             self.initiators[key] = NVMeInitiator(node, self.gateways[0], nqn)
 
         return self.initiators[key]
-
-    def create_dhchap_key(self, config, update_host_key=False):
-        """Generate DHCHAP key for each initiator and store it."""
-        subnqn = config["subnqn"]
-        group = config["gw_group"]
-        nqn = f"{subnqn}.{group}"
-
-        for host_config in config["hosts"]:
-            node_id = host_config["node"]
-            initiator = self.get_or_create_initiator(node_id, nqn)
-
-            # Generate key for subsystem NQN
-            key, _ = initiator.gen_dhchap_key(n=config["subnqn"])
-            LOG.info(f"{key.strip()} is generated for {nqn} and {node_id}")
-
-            initiator.nqn = config["subnqn"]
-            initiator.auth_mode = config.get("auth_mode")
-            if initiator.auth_mode == "bidirectional" and not update_host_key:
-                initiator.subsys_key = key.strip()
-                initiator.host_key = key.strip()
-            if initiator.auth_mode == "unidirectional":
-                initiator.host_key = key.strip()
-            if update_host_key:
-                initiator.host_key = key.strip()
-            config["dhchap-key"] = key.strip()
-
-            self.clients.append(initiator)
 
     def catogorize(self, gws):
         """Categorize to-be failed and running GWs.
@@ -136,66 +87,6 @@ class HighAvailability:
                 running_gws.append(gw)
 
         return fail_gws, running_gws
-
-    @staticmethod
-    def string_to_dict(string):
-        """Parse ANA states from the string."""
-        states = string.replace(" ", "").split(",")
-        dict = {}
-        for state in states:
-            if not state:
-                continue
-            _id, _state = state.split(":")
-            dict[int(_id)] = _state
-        return dict
-
-    def ana_states(self, gw_group=""):
-        """Fetch ANA states and convert into python dict."""
-
-        out, _ = self.orch.shell(
-            args=["ceph", "nvme-gw", "show", self.nvme_pool, repr(self.gateway_group)]
-        )
-        states = {}
-        if self.cluster.rhcs_version >= "8":
-            out = json.loads(out)
-            for gateway in out.get("Created Gateways:"):
-                gw = gateway["gw-id"]
-                states[gw] = gateway
-                states[gw].update(self.string_to_dict(gateway["ana states"]))
-        else:
-            for data in out.split("}"):
-                data = data.strip()
-                if not data:
-                    continue
-                data = json.loads(f"{data}}}")
-                if data.get("ana states"):
-                    gw = data["gw-id"]
-                    states[gw] = data
-                    states[gw].update(self.string_to_dict(data["ana states"]))
-
-        return states
-
-    def check_gateway_availability(self, ana_id, state="AVAILABLE", ana_states=None):
-        """Check for failed ANA GW become unavailable.
-
-        Args:
-            ana_id: Gateway ANA group id.
-            state: Gateway availability state
-            ana_states: Overall ana state. (output from self.ana_states)
-        Return:
-            True if Gateway availability is in expected state, else False
-        """
-        # get ANA states
-        if not ana_states:
-            ana_states = self.ana_states()
-
-        # Check Availability of ANA Group Gateway
-        for _, _state in ana_states.items():
-            if _state["anagrp-id"] == ana_id:
-                if _state["Availability"] == state:
-                    return True
-                return False
-        return False
 
     def get_optimized_state(self, failed_ana_id):
         """Fetch the Optimized ANA states for failed gateway.
@@ -499,246 +390,6 @@ class HighAvailability:
 
         return False
 
-    def scale_down(self, gateway_nodes_to_be_scaleddown):
-        """Scaling down of the NVMeoF Gateways.
-
-        Initiate scale-down
-        - List the gateways which has to be scaled down.
-        - Validate the ANA states of scaled down GWs are optimized in one of the other working GWs.
-
-        Post scale-down Validation
-        - List out namespaces associated with the scaled down Gateways using ANA group ids.
-        - Check for 5 Consecutive times for the increments in write/read to validate IO continuation.
-        """
-        start_counter = float()
-        start_time = str()
-        end_counter = float()
-        end_time = str()
-        LOG.info(f"{gateway_nodes_to_be_scaleddown}: Scaling down NVMe Service")
-
-        if not isinstance(gateway_nodes_to_be_scaleddown, list):
-            gateway_nodes_to_be_scaleddown = [gateway_nodes_to_be_scaleddown]
-
-        to_be_scaledown_gws, operational_gws = self.catogorize(
-            gateway_nodes_to_be_scaleddown
-        )
-        ana_ids = [gw.ana_group_id for gw in to_be_scaledown_gws]
-        gateway = operational_gws[0]
-
-        # Validate IO and scale operation
-        old_namespaces = self.fetch_namespaces(gateway, ana_ids)
-        self.validate_io(old_namespaces)
-
-        # Scale down
-        gwnodes_to_be_deployed = list(
-            set(self.config["gw_nodes"]) - set(gateway_nodes_to_be_scaleddown)
-        )
-        self.config["gw_nodes"] = gwnodes_to_be_deployed
-        deploy_nvme_service(self.cluster, self.config)
-
-        self.gateways = []
-        for gateway in self.config["gw_nodes"]:
-            gw_node = get_node_by_id(self.cluster, gateway)
-            self.gateways.append(NVMeGateway(gw_node, self.mtls))
-
-        start_counter, start_time = get_current_timestamp()
-        for gateway in to_be_scaledown_gws:
-            hostname = gateway.hostname
-
-            if self.cluster.rhcs_version == "8":
-                # Wait until 60 seconds
-                for w in WaitUntil():
-                    # Check for gateway unavailability
-                    if self.check_gateway_availability(
-                        gateway.ana_group_id, state="DELETING"
-                    ):
-                        LOG.info(f"[ {gateway} ] NVMeofGW service is UNAVAILABLE.")
-                        active = self.get_optimized_state(gateway.ana_group_id)
-
-                        # Find optimized path
-                        if active:
-                            LOG.info(
-                                f"{list(active[0])} is new and only Active GW for failed {hostname}"
-                            )
-                            break
-
-                    LOG.warning(f"[ {hostname} ] is still in AVAILABLE state..")
-
-                if w.expired:
-                    raise TimeoutError(
-                        f"[ {hostname} ] Scale down of NVMeofGW service failed after 60s timeout.."
-                    )
-
-            end_counter, end_time = get_current_timestamp()
-            LOG.info(
-                f"[ {hostname} ] Total time taken to scale down - {end_counter - start_counter} seconds"
-            )
-
-            result = {
-                "scale-down-start-time": start_time,
-                "scale-down-end-time": end_time,
-                "scale-down-start-counter-time": start_counter,
-                "scale-down-end-counter-time": end_counter,
-            }
-            LOG.info(log_json_dump(result))
-
-            # Validate auto load balance if rhcs version is 8.1
-            if self.cluster.rhcs_version == "8.1":
-                time.sleep(60)
-                validate_ns_balance = self.validate_auto_loadbalance()
-                LOG.info(f"Validated namespaces in each GW:{validate_ns_balance}")
-            # Validate IO post scale down
-            self.validate_io(set(list(old_namespaces)))
-            return result
-
-    def validate_scaleup(self, scaleup_nodes, namespaces):
-        """
-        - List out namespaces associated with the new Gateways using ANA group ids.
-        - Check for 5 Consecutive times for the increments in write/read to validate IO continuation.
-
-        Args:
-            scaleup_nodes (list): A list of gateway nodes to be scaled up.
-        """
-        start_counter = float()
-        start_time = str()
-        end_counter = float()
-        end_time = str()
-        new_gws = []
-
-        for gateway_node in scaleup_nodes:
-            gw = get_node_by_id(self.cluster, gateway_node)
-            new_gws.append(NVMeGateway(gw, self.mtls))
-
-        start_counter, start_time = get_current_timestamp()
-        for gateway in new_gws:
-            hostname = gateway.hostname
-
-            # Wait until 60 seconds
-            for w in WaitUntil(timeout=60):
-                # Check for gateway availability
-                if self.check_gateway_availability(gateway.ana_group_id):
-                    LOG.info(f"[ {gateway} ] NVMeofGW service is AVAILABLE.")
-                    state = self.get_optimized_state(gateway.ana_group_id)
-
-                    # check gateway for its own original path.
-                    if gateway.ana_group["name"] in state[0]:
-                        end_counter, end_time = get_current_timestamp()
-                        LOG.info(
-                            f"{hostname} restored to original path - {log_json_dump(state)}"
-                        )
-                        break
-
-                LOG.warning(f"[ {hostname} ] is still not in AVAILABLE state..")
-
-            if w.expired:
-                raise TimeoutError(
-                    f"[ {hostname} ] Scale up of NVMeofGW service failed after 120s timeout.."
-                )
-
-            LOG.info(
-                f"[ {hostname} ] Total time taken to scale up - {end_counter - start_counter} seconds"
-            )
-            result = {
-                "scale-up-start-time": start_time,
-                "scale-up-end-time": end_time,
-                "scale-up-start-counter-time": start_counter,
-                "scale-up-end-counter-time": end_counter,
-            }
-            LOG.info(log_json_dump(result))
-            # Validate auto load balance if rhcs version is 8.1
-            if self.cluster.rhcs_version == "8.1":
-                time.sleep(60)
-                validate_ns_balance = self.validate_auto_loadbalance()
-                LOG.info(f"Validated namespaces in each GW:{validate_ns_balance}")
-            # Validate IO post scale up
-            self.validate_io(set(list(namespaces)))
-            return result
-
-    def scale_up(self, scaleup_nodes, gw_nodes, existing_namespaces):
-        """Scaling up of the NVMeoF Gateways.
-
-        Initiate scale-up
-        - Spin up the new gateways.
-        - Validate the ANA states of new GWs are optimized.
-
-        Pre scale-up Validation
-        - List out namespaces associated with the new Gateways using ANA group ids.
-        - Check for 5 Consecutive times for the increments in write/read to validate IO continuation.
-
-        Post scale-up Validation
-        - Check if Ana group ids of replaced GWs took over the original ANA group ids
-        """
-        existing_namespaces = []
-        LOG.info(f"{scaleup_nodes}: Scaling up NVMe Service")
-
-        if not isinstance(scaleup_nodes, list):
-            scaleup_nodes = [scaleup_nodes]
-
-        # Validate IO before scale up operation
-        self.validate_io(set(list(existing_namespaces)))
-
-        # Scale up
-        gwnodes_to_be_deployed = list(set(self.config["gw_nodes"] + scaleup_nodes))
-        self.config["gw_nodes"] = gwnodes_to_be_deployed
-        deploy_nvme_service(self.cluster, self.config)
-
-        self.gateways = []
-        for gateway in gwnodes_to_be_deployed:
-            gw_node = get_node_by_id(self.cluster, gateway)
-            self.gateways.append(NVMeGateway(gw_node, self.mtls))
-
-        # Validate ana_grp_ids post scale up
-        for scaleup_node in scaleup_nodes:
-            for gw_node in gw_nodes:
-                if gw_node.node.id == scaleup_node:
-                    gw = self.check_gateway(gw_node.node.id)
-                    scaleup_gw = self.check_gateway(scaleup_node)
-                    if gw.ana_group_id == scaleup_gw.ana_group_id:
-                        LOG.info("Scaleup nodes took over the previous anagrpids")
-                    else:
-                        raise Exception("anagrpids are not matching after scaleup")
-
-    def validate_auto_loadbalance(self, gw_group=""):
-        """
-        Fetch the namespace count on each Gateway and compare them.
-        Ensure that the number of namespaces for each GW is within the range [num_namespaces_per_gw + or - len(GWs)].
-        """
-        out, _ = self.orch.shell(
-            args=["ceph", "nvme-gw", "show", self.nvme_pool, repr(self.gateway_group)]
-        )
-        out = json.loads(out)
-        total_num_namespaces = out.get("num-namespaces")
-        gateways = out.get("Created Gateways:", [])
-        total_gateways = len(gateways)
-        if total_gateways == 0:
-            raise Exception("No gateways found in the output.")
-
-        num_namespaces_per_gw = total_num_namespaces / total_gateways
-        namespaces = {}
-        LOG.info(f"Total namespace in GW group : {total_num_namespaces}")
-        LOG.info(f"Total GWs: {total_gateways}")
-        LOG.info(f"Namespaces per GW : {num_namespaces_per_gw}")
-
-        for gateway in gateways:
-            gw_id = gateway["gw-id"]
-            num_namespaces = gateway["num-namespaces"]
-            lower_range = num_namespaces_per_gw - total_gateways
-            upper_range = num_namespaces_per_gw + total_gateways
-            LOG.info(
-                f"namespace per GW must be in range between {lower_range} and {upper_range}"
-            )
-
-            if not (lower_range <= num_namespaces <= upper_range):
-                raise Exception(
-                    f"Gateway '{gw_id}' has an invalid num-namespaces: {num_namespaces}. "
-                    f"It must be between {lower_range} and {upper_range}."
-                )
-
-            namespaces[gw_id] = gateway
-            namespaces[gw_id]["num-namespaces"] = num_namespaces
-
-        return namespaces
-
     def failover(self, gateway, fail_tool):
         """HA Failover on the NVMeoF Gateways.
 
@@ -1010,257 +661,6 @@ class HighAvailability:
                 LOG.info(f"IO validation for {subsys}|{pool_img} is successful.")
 
         LOG.info("IO Validation is Successfull on all RBD images..")
-
-    def validate_init_namespace_masking(
-        self,
-        command,
-        init_nodes,
-        expected_visibility,
-        validate_config=None,
-    ):
-        """Validate that the namespace visibility is correct from all initiators."""
-        for node in init_nodes:
-            initiator_node = get_node_by_id(self.cluster, node)
-            client = NVMeInitiator(initiator_node, self.gateways[0])
-            client.disconnect_all()  # Reconnect NVMe targets
-            client.connect_targets(config={"nqn": "connect-all"})
-            serial_to_namespace = defaultdict(set)
-
-            out, _ = initiator_node.exec_command(
-                sudo=True, cmd="cat /etc/os-release | grep VERSION_ID"
-            )
-            rhel_version = out.split("=")[1].strip().strip('"')
-
-            @retry(
-                IOError,
-                tries=4,
-                delay=3,
-            )
-            def execute_nvme_command(client_node):
-                devices_json, _ = client_node.exec_command(
-                    cmd="nvme list --output-format=json", sudo=True
-                )
-                return json.loads(devices_json)["Devices"]
-
-            devices_json = execute_nvme_command(initiator_node)
-            if not devices_json:
-                LOG.info(f"No devices found on node {node}")
-                continue
-
-            for device in devices_json:
-                if rhel_version == "9.5":
-                    key = device["NameSpace"]
-                    value = int(device["SerialNumber"])
-                    serial_to_namespace[key].add(value)
-                elif rhel_version == "9.6":
-                    for subsys in device.get("Subsystems", []):
-                        for controller in subsys.get("Controllers", []):
-                            if controller.get("ModelNumber") == "Ceph bdev Controller":
-                                serial = controller.get("SerialNumber", "")
-                                value = int(serial)
-                                for ns in subsys.get("Namespaces", []):
-                                    key = ns.get("NSID")
-                                    serial_to_namespace[key].add(value)
-
-            def subsystem_nsid_found(dictionary, key, value):
-                return key in dictionary and value in dictionary[key]
-
-            if validate_config:
-                args = (validate_config or {}).get("args", {})
-                subsystem_to_nsid = {args["nsid"]: args["sub_num"]}
-                init_node = args.get("init_node")
-                ns_to_check, subsystem_to_check = next(iter(subsystem_to_nsid.items()))
-                LOG.info(f"{subsystem_to_nsid} : {serial_to_namespace}")
-                ns_subsys_found = subsystem_nsid_found(
-                    serial_to_namespace, ns_to_check, subsystem_to_check
-                )
-
-                if command == "add_host":
-                    if node == init_node:
-                        if ns_subsys_found:
-                            LOG.info(
-                                f"Validated - Namespace:Subsystem pair {subsystem_to_nsid} is listed on {node}"
-                            )
-                        else:
-                            LOG.error(
-                                f"Namespace:Subsystem pair {subsystem_to_nsid} is not listed on {node}"
-                            )
-                            raise Exception(
-                                f"Expected Namespace:Subsystem pair {subsystem_to_nsid} on {node} but did not find it"
-                            )
-                    else:
-                        if ns_subsys_found:
-                            LOG.error(
-                                f"Namespace:Subsystem pair {subsystem_to_nsid} is listed on {node}"
-                            )
-                            raise Exception(
-                                f"Did not expect Namespace:Subsystem pair {subsystem_to_nsid} on {node} but found it"
-                            )
-                        else:
-                            LOG.info(
-                                f"Validated - Namespace:Subsystem pair {subsystem_to_nsid} is not listed on {node}"
-                            )
-                elif command == "del_host":
-                    if node == init_node:
-                        if ns_subsys_found:
-                            LOG.error(
-                                f"Namespace:Subsystem pair {subsystem_to_nsid} is listed on {node}"
-                            )
-                            raise Exception(
-                                f"Did not expect Namespace:Subsystem pair {subsystem_to_nsid} on {node} but found it"
-                            )
-                        else:
-                            LOG.info(
-                                f"Validated - Namespace:Subsystem pair {subsystem_to_nsid} is not listed on {node}"
-                            )
-                    else:
-                        if ns_subsys_found:
-                            LOG.error(
-                                f"Namespace:Subsystem pair {subsystem_to_nsid} is listed on {node}"
-                            )
-                            raise Exception(
-                                f"Did not expect Namespace:Subsystem pair {subsystem_to_nsid} on {node} but found it"
-                            )
-                        else:
-                            LOG.info(
-                                f"Validated - Namespace:Subsystem pair {subsystem_to_nsid} is not listed on {node}"
-                            )
-            else:
-                if (
-                    not expected_visibility
-                ):  # If expected visibility is False, devices should be empty
-                    # Determine if devices list is empty (no Namespaces in any Subsystem)
-                    devices_json_empty = (
-                        all(
-                            not subsys.get("Namespaces")  # True if empty or missing
-                            for device in devices_json
-                            for subsys in device.get("Subsystems", [])
-                        )
-                        if rhel_version == "9.6"
-                        else not devices_json
-                    )
-                    if not devices_json_empty:  # Check if Devices is not empty
-                        LOG.error(
-                            f"Expected no devices for initiator {node}, but found: {devices_json}"
-                        )
-                        raise Exception(
-                            f"Initiator {node} has devices when NS visibility is restricted"
-                        )
-                    else:
-                        LOG.info(f"Validated - no devices found on {node}")
-                elif (
-                    expected_visibility
-                ):  # If expected visibility is True, devices should not be empty
-                    devices_json_empty = (
-                        all(
-                            not subsys.get("Namespaces")  # True if empty or missing
-                            for device in devices_json
-                            for subsys in device.get("Subsystems", [])
-                        )
-                        if rhel_version == "9.6"
-                        else not devices_json
-                    )
-                    if devices_json_empty:
-                        LOG.error(
-                            f"Expected devices to be visible for node {node}, but found none."
-                        )
-                        raise Exception(
-                            f"Initiator {node} has no devices when NS visibility is restricted"
-                        )
-                    else:
-                        # Log Namespace and SerialNumber from each device
-                        for device in devices_json:
-                            if rhel_version == "9.5":
-                                namespace = device.get("NameSpace", None)
-                                serial_number = device.get("SerialNumber", None)
-                            elif rhel_version == "9.6":
-                                for subsys in device.get("Subsystems", []):
-                                    for controller in subsys.get("Controllers", []):
-                                        if (
-                                            controller.get("ModelNumber")
-                                            == "Ceph bdev Controller"
-                                        ):
-                                            serial_number = int(
-                                                controller.get("SerialNumber", "")
-                                            )
-                                            for ns in subsys.get("Namespaces", []):
-                                                namespace = ns.get("NSID")
-                                                LOG.info(
-                                                    f"Namespace: {namespace}, SerialNumber: {serial_number}"
-                                                )
-                        LOG.info(
-                            f"Validated - {len(devices_json)} devices found on {node}"
-                        )
-
-    def validate_namespace_masking(
-        self,
-        nsid,
-        subnqn,
-        namespaces_sub,
-        hostnqn_dict,
-        ns_visibility,
-        command,
-        expected_visibility,
-    ):
-        """Validate that the namespace visibility is correct."""
-
-        if command == "add_host":
-            LOG.info(command)
-            num_namespaces_per_node = namespaces_sub // len(hostnqn_dict)
-
-            # Determine the initiator node responsible for this nsid based on the calculated range
-            node_index = (nsid - 1) // num_namespaces_per_node
-            expected_host = list(hostnqn_dict.values())[node_index]
-
-            # Log the visibility of the namespace
-            if expected_host in ns_visibility:
-                LOG.info(
-                    f"Validated - Namespace {nsid} of {subnqn} has the correct nqn {ns_visibility}"
-                )
-            else:
-                LOG.error(
-                    f"Namespace {nsid} of {subnqn} has incorrect NQN. Expected {expected_host}, but got {ns_visibility}"
-                )
-                raise Exception(
-                    f"Namespace {nsid} of {subnqn} has incorrect NQN. Expected {expected_host}, but got {ns_visibility}"
-                )
-
-        elif command == "del_host":
-            LOG.info(command)
-            num_namespaces_per_node = namespaces_sub // len(hostnqn_dict)
-
-            # Determine the initiator node responsible for this nsid based on the calculated range
-            node_index = (nsid - 1) // num_namespaces_per_node
-            expected_host = list(hostnqn_dict.values())[node_index]
-
-            # Log the visibility of the namespace
-            if expected_host not in ns_visibility:
-                LOG.info(
-                    f"Validated - Namespace {nsid} of {subnqn} does not has {expected_host}"
-                )
-            else:
-                LOG.error(
-                    f"Namespace {nsid} of {subnqn} has {ns_visibility} which was removed"
-                )
-                raise Exception(
-                    f"Namespace {nsid} of {subnqn} has incorrect NQN. Not expecting {ns_visibility} in {expected_host}"
-                )
-
-        else:
-            # Validate visibility based on the expected value (for non-add/del host commands)
-            # ns_visibility = str(ns_visibility)
-            LOG.info(command)
-            # if ns_visibility.lower() == expected_visibility.lower():
-            if ns_visibility == expected_visibility:
-                LOG.info(
-                    f"Validated - Namespace {nsid} has correct visibility: {ns_visibility}"
-                )
-                LOG.error(
-                    f"NS {nsid} of {subnqn} has wrong visibility.Expected {expected_visibility} got{ns_visibility}"
-                )
-                raise Exception(
-                    f"NS {nsid} of {subnqn} has wrong visibility.Expected {expected_visibility} got {ns_visibility}"
-                )
 
     def run(self):
         """Execute the HA failover and failback with IO validation."""
